# Speed up cellular noise